Move jieba and spacy wrapper to a pipeline.external module
